frontend/requirements.txt
streamlit==1.29.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.3
plotly==5.18.0
markdown==3.5.1 